add inserting package name functionality
Adds insert package functionality:

While this is not a core feature of toggling tests, it is somewhat useful
when creating new classes. Given that all the necessary config and supporting
functions exist to implement this feature, I went ahead and created it.

Fixes numerous issues with newer Sublime Text versions:

* The command and context menus where not displaying due to a change in a constant
  used by Sublime Text 3.
* Scoggle commands are now visible in the context menu.